move within cache
